feat: create nextjs app and template home page
