+++ .prettierrc
@@ -1,0 +1,3 @@
+{
+	"useTabs": true
+}
