+++ apps/site/src/app/page.tsx
@@ -1,0 +1,1 @@
+export { Home as default } from "./views";
