+++ apps/site/src/app/views/Home/Home.tsx
@@ -1,0 +1,3 @@
+export default function Home() {
+	return <h1 className="text-3xl font-bold underline">IrvineHacks 2024</h1>;
+}
