+++ apps/site/src/app/views/index.ts
@@ -1,0 +1,1 @@
+export { default as Home } from "./Home/Home";
