--- apps/web/next.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  reactStrictMode: true,
-  transpilePackages: ["ui"],
-};
